# nb2/models.py
from nb2 import db


class Person(db.Model):
    """
    Represents someone that Nostalgiabot has quotes for;
    usually an SDE employee.
    """
    id = db.Column(db.Integer, primary_key=True)
    slack_user_id = db.Column(
        db.String(16),
        index=True,
        unique=True,
        nullable=False
    )
    first_name = db.Column(db.String(32), nullable=False)
    last_name = db.Column(db.String(32))
    quotes = db.relationship('Quote', backref='person', lazy=True)

    required_fields = ['slack_user_id', 'first_name']
    editable_fields = ['slack_user_id', 'first_name', 'last_name']
    include_fields = ['quotes']

    def __repr__(self):
        return f"<Person: {self.slack_user_id} |  Name: {self.first_name} | Id: {self.id}>"

    @property
    def get_quotes(self):
        return [quote.content for quote in self.quotes]

    def serialize(self, includes=None):
        includes = includes or []
        data = {
            'id': self.id,
            'slack_user_id': self.slack_user_id,
            'first_name': self.first_name,
            'last_name': self.last_name
        }

        for field in set(includes) & set(self.include_fields):
            data.update({field: getattr(self, f"get_{field}")})

        return data

    def deserialize(self, data):
        """
        Update the editable fields on a person with `data`.
        """
        for field in self.editable_fields:
            if field in data:
                setattr(self, field, data[field])


class Quote(db.Model):
    """
    Represents something a Person has said;
    usually something funny.
    """
    id = db.Column(db.Integer, primary_key=True)
    content = db.Column(db.Text, nullable=False)
    person_id = db.Column(
        db.Integer,
        db.ForeignKey('person.id'),
        nullable=False
    )
    created = db.Column(db.DateTime, nullable=False)

    def __repr__(self):
        return f"<Quote: {self.content} | Id: {self.id}>"